pyproject.toml
[project]
name = "moonfish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "apsw>=3.49.1.0",
    "cryptography>=44.0.2",
    "fastapi[standard]>=0.115.11",
    "google-genai>=1.8.0",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "sqlmodel>=0.0.24",
    "sse-starlette>=2.2.1",
]

[dependency-groups]
dev = ["ipykernel>=6.29.5"]
